requirements.txt
annotated-types==0.6.0
anyio==4.3.0
fastapi==0.110.2
icmplib==3.0.4
idna==3.7
netifaces==0.11.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
